[OCISDEV-628] Test renamed
Test renamed from `130_ldap_group_writes` to `130-ldap-create-group-add-users-and-delete-group`
